anvil_backend/requirements.txt
Pillow
requests
anvil-uplink
python-dotenv